refact: controleur
