+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "Probe-api",
-  "version": "0.3.88",
+  "version": "0.3.89",
   "description": "API pour récupérer des informations sur l'application Probe",
   "main": "app.js",
   "scripts": {
